addAwesomeMarkers {leaflet}	R Documentation

Add Awesome Markers

Description

Add Awesome Markers

Usage

addAwesomeMarkers(map, lng = NULL, lat = NULL, layerId = NULL,
  group = NULL, icon = NULL, popup = NULL, popupOptions = NULL,
  label = NULL, labelOptions = NULL, options = markerOptions(),
  clusterOptions = NULL, clusterId = NULL, data = getMapData(map))

Arguments

map
the map to add awesome Markers to.

lng
a numeric vector of longitudes, or a one-sided formula of the form ~x where x is a variable in data; by default (if not explicitly provided), it will be automatically inferred from data by looking for a column named lng, long, or longitude (case-insensitively)

lat
a vector of latitudes or a formula (similar to the lng argument; the names lat and latitude are used when guessing the latitude column from data)

layerId
the layer id

group
the name of the group the newly created layers should belong to (for clearGroup and addLayersControl purposes). Human-friendly group names are permitted–they need not be short, identifier-style names. Any number of layers and even different types of layers (e.g. markers and polygons) can share the same group name.

icon
the icon(s) for markers;

popup
a character vector of the HTML content for the popups (you are recommended to escape the text using htmlEscape()

popupOptions
A Vector of popupOptions to provide popups for security reasons)

label
a character vector of the HTML content for the labels

labelOptions
A Vector of labelOptions to provide label options for each label. Default NULL

options
a list of extra options for tile layers, popups, paths (circles, rectangles, polygons, ...), or other map elements

clusterOptions
if not NULL, markers will be clustered using Leaflet.markercluster; you can use markerClusterOptions() to specify marker cluster options

clusterId
the id for the marker cluster layer

data
the data object from which the argument values are derived; by default, it is the data object provided to leaflet() initially, but can be overridden

[Package leaflet version 2.0.1 Index]