addProviderTiles {leaflet}	R Documentation

Add a tile layer from a known map provider

Description

Add a tile layer from a known map provider

Usage

addProviderTiles(map, provider, layerId = NULL, group = NULL,
  options = providerTileOptions())

providerTileOptions(errorTileUrl = "", noWrap = FALSE, opacity = NULL,
  zIndex = NULL, updateWhenIdle = NULL, detectRetina = FALSE, ...)

Arguments

map
the map to add the tile layer to

provider
the name of the provider (see http://leaflet-extras.github.io/leaflet-providers/preview/ and https://github.com/leaflet-extras/leaflet-providers)

layerId
the layer id to assign

group
the name of the group the newly created layers should belong to (for clearGroup and addLayersControl purposes). Human-friendly group names are permitted–they need not be short, identifier-style names.

options
tile options

errorTileUrl, noWrap, opacity, zIndex, updateWhenIdle, detectRetina
the tile layer options; see http://leafletjs.com/reference-1.3.1.html#tilelayer

...
named parameters to add to the options

Value

modified map object

Examples

leaflet() %>%
  addProviderTiles("Stamen.Watercolor") %>%
  addProviderTiles("Stamen.TonerHybrid")

[Package leaflet version 2.0.1 Index]