addTerminator {leaflet}	R Documentation

Add a daylight layer on top of the map

Description

See https://github.com/joergdietrich/Leaflet.Terminator

Usage

addTerminator(map, resolution = 2, time = NULL, layerId = NULL,
  group = NULL, options = pathOptions(pointerEvents = "none", clickable =
  FALSE))

Arguments

map
a map widget object

resolution
the step size at which the terminator points are computed. The step size is 1 degree/resolution, i.e. higher resolution values have smaller step sizes and more points in the polygon. The default value is 2.

time
Time

layerId
the layer id

group
the name of the group this layer belongs to.

options
the path options for the daynight layer

Examples

leaf <- leaflet() %>%
  addTiles() %>%
  addTerminator()
leaf

[Package leaflet version 2.0.0 Index]