Create a list of awesome icon data see https://github.com/lennardv2/Leaflet.awesome-markers

An icon can be represented as a list of the form list(icon, library, ...). This function is vectorized over its arguments to create a list of icon data. Shorter argument values will be re-cycled. NULL values for these arguments will be ignored.

Usage

awesomeIcons(
  icon = "home",
  library = "glyphicon",
  markerColor = "blue",
  iconColor = "white",
  spin = FALSE,
  extraClasses = NULL,
  squareMarker = FALSE,
  iconRotate = 0,
  fontFamily = "monospace",
  text = NULL
)

Arguments

icon

Name of the icon

library

Which icon library. Default "glyphicon", other possible values are "fa" (fontawesome) or "ion" (ionicons).

markerColor

Possible values are "red", "darkred", "lightred", "orange", "beige", "green", "darkgreen", "lightgreen", "blue", "darkblue", "lightblue", "purple", "darkpurple", "pink", "cadetblue", "white", "gray", "lightgray", "black"

iconColor

The color to use for the icon itself. Use any CSS-valid color (hex, rgba, etc.) or a named web color.

spin

If TRUE, make the icon spin (only works when library = "fa")

extraClasses

Additional css classes to include on the icon.

squareMarker

Whether to use a square marker.

iconRotate

Rotate the icon by a given angle.

fontFamily

Used when text option is specified.

text

Use this text string instead of an icon. argument of addAwesomeMarkers().

Value

A list of awesome-icon data that can be passed to the icon