Normalize data to a data frame or Spatial object.

Given a data object and lng/lat arguments (which may be NULL) [meaning infer from data], formula [which should be evaluated with respect to the data], or vector data [which should be used as-is]) return a lng/lat data frame.

Usage

derivePoints(
  data,
  lng = NULL,
  lat = NULL,
  missingLng = missing(lng),
  missingLat = missing(lat),
  funcName = "f"
)

derivePolygons(
  data,
  lng = NULL,
  lat = NULL,
  missingLng = missing(lng),
  missingLat = missing(lat),
  funcName = "f"
)

Arguments

data

map data

lng, lat

longitude and latitude

missingLng

whether lng is missing

missingLat

whether lat is missing

funcName

Name of calling function (for logging)

Value

derivePoints() returns a lng/lat data frame.

derivePolygons() returns a list with bbox attributes.