derivePolygons {leaflet}	R Documentation

Given a data object and lng/lat arguments (which may be NULL [meaning infer from data], formula [which should be evaluated with respect to the data], or vector data [which should be used as-is]) return a spatial object

Description

Given a data object and lng/lat arguments (which may be NULL [meaning infer from data], formula [which should be evaluated with respect to the data], or vector data [which should be used as-is]) return a spatial object

Usage

derivePolygons(data, lng = NULL, lat = NULL, missingLng = missing(lng),
  missingLat = missing(lat), funcName = "f")

Arguments

data
map data

lng
longitude

lat
latitude

missingLng
whether lng is missing

missingLat
whether lat is missing

funcName
Name of calling function (for logging)

[Package leaflet version 2.0.1 Index]