dispatch {leaflet}	R Documentation

Extension points for plugins

Description

Extension points for plugins

Usage

dispatch(map, funcName, leaflet = stop(paste(funcName,
  "requires a map proxy object")), leaflet_proxy = stop(paste(funcName,
  "does not support map proxy objects")))

invokeMethod(map, data, method, ...)

Arguments

map
a map object, as returned from leaflet or leafletProxy

funcName
the name of the function that the user called that caused this dispatch call; for error message purposes

leaflet
an action to be performed if the map is from leaflet

leaflet_proxy
an action to be performed if the map is from leafletProxy

data
a data object that will be used when evaluating formulas in ...

method
the name of the JavaScript method to invoke

...
unnamed arguments to be passed to the JavaScript method

Value

dispatch returns the value of leaflet or leaflet_proxy, or an error. invokeMethod returns the map object that was passed in, possibly modified.

[Package leaflet version 2.0.1 Index]