evalFormula {leaflet}	R Documentation

Evaluate list members that are formulae, using the map data as the environment (if provided, otherwise the formula environment)

Description

Evaluate list members that are formulae, using the map data as the environment (if provided, otherwise the formula environment)

Usage

evalFormula(list, data)

Arguments

list
with members as formulae

data
map data

[Package leaflet version 2.0.3 Index]