leafletDependencies {leaflet}	R Documentation

Various leaflet dependency functions for use in downstream packages

Description

Various leaflet dependency functions for use in downstream packages

Usage

leafletDependencies

Format

An object of class list of length 13.

Examples

## Not run: 
addBootStrap <- function(map) {
  map$dependencies <- c(map$dependencies, leafletDependencies$bootstrap())
  map
}

## End(Not run)

[Package leaflet version 2.0.3 Index]