Leaflet sizing policy

Sizing policy used withing leaflet htmlwidgets.

Usage

leafletSizingPolicy(
  defaultWidth = "100%",
  defaultHeight = 400,
  padding = 0,
  browser.fill = TRUE,
  ...
)

Arguments

defaultWidth

defaults to "100%" of the available width

defaultHeight

defaults to 400px tall

padding

defaults to 0px

browser.fill

defaults to TRUE

...

Arguments passed on to htmlwidgets::sizingPolicy

viewer.defaultWidth

The default width used to display the widget within the RStudio Viewer.

viewer.defaultHeight

The default height used to display the widget within the RStudio Viewer.

viewer.padding

Padding around the widget when displayed in the RStudio Viewer (defaults to 15 pixels).

viewer.fill

When displayed in the RStudio Viewer, automatically size the widget to the viewer dimensions (note that viewer.padding is still applied). Default to TRUE.

viewer.suppress

Never display the widget within the RStudio Viewer (useful for widgets that require a large amount of space for rendering). Defaults to FALSE.

viewer.paneHeight

Request that the RStudio Viewer be forced to a specific height when displaying this widget.

browser.defaultWidth

The default width used to display the widget within a standalone web browser.

browser.defaultHeight

The default height used to display the widget within a standalone web browser.

browser.padding

Padding around the widget when displayed in a standalone browser (defaults to 40 pixels).

browser.external

When displaying in a browser, always use an external browser (via browseURL()). Defaults to `FALSE“, which will result in the use of an internal browser within RStudio v1.1 and higher.

knitr.defaultWidth

The default width used to display the widget within documents generated by knitr (e.g. R Markdown).

knitr.defaultHeight

The default height used to display the widget within documents generated by knitr (e.g. R Markdown).

knitr.figure

Apply the default knitr fig.width and fig.height to the widget when it's rendered within R Markdown documents. Defaults to TRUE.

fill

Whether or not the widget's container should be treated as a fill item, meaning that its height is allowed to grow/shrink to fit a fill container with an opinionated height (see htmltools::bindFillRole() for more). Examples of fill containers include bslib::card() and bslib::card_body_fill().

Value

An htmlwidgets::sizingPolicy object