makeIcon {leaflet}	R Documentation

Define icon sets

Description

Define icon sets

Usage

makeIcon(iconUrl = NULL, iconRetinaUrl = NULL, iconWidth = NULL,
  iconHeight = NULL, iconAnchorX = NULL, iconAnchorY = NULL,
  shadowUrl = NULL, shadowRetinaUrl = NULL, shadowWidth = NULL,
  shadowHeight = NULL, shadowAnchorX = NULL, shadowAnchorY = NULL,
  popupAnchorX = NULL, popupAnchorY = NULL, className = NULL)

Arguments

iconUrl
the URL or file path to the icon image

iconRetinaUrl
the URL or file path to a retina sized version of the icon image

iconWidth
size of the icon image in pixels

iconHeight
size of the icon image in pixels

iconAnchorX
the coordinates of the "tip" of the icon (relative to its top left corner, i.e. the top left corner means iconAnchorX = 0 and iconAnchorY = 0), and the icon will be aligned so that this point is at the marker's geographical location

iconAnchorY
the coordinates of the "tip" of the icon (relative to its top left corner, i.e. the top left corner means iconAnchorX = 0 and iconAnchorY = 0), and the icon will be aligned so that this point is at the marker's geographical location

shadowUrl
the URL or file path to the icon shadow image

shadowRetinaUrl
the URL or file path to the retina sized version of the icon shadow image

shadowWidth
size of the shadow image in pixels

shadowHeight
size of the shadow image in pixels

shadowAnchorX
the coordinates of the "tip" of the shadow

shadowAnchorY
the coordinates of the "tip" of the shadow

popupAnchorX
the coordinates of the point from which popups will "open", relative to the icon anchor

popupAnchorY
the coordinates of the point from which popups will "open", relative to the icon anchor

className
a custom class name to assign to both icon and shadow images

[Package leaflet version 2.0.0 Index]