safeLabel {leaflet}	R Documentation

Sanitize textual labels

Description

This is a helper function used internally to HTML-escape user-provided labels. It prevents strings from unintentionally being treated as HTML when they are intended to be plaintext.

Usage

safeLabel(label, data)

Arguments

label
A vector or list of plain characters or HTML (marked by HTML), or a formula that resolves to such a value.

data
A data frame over which the formula is evaluated.

[Package leaflet version 2.0.1 Index]