showGroup {leaflet}	R Documentation

Show or hide layer groups

Description

Hide groups of layers without removing them from the map entirely. Groups are created using the group parameter that is included on most layer adding functions.

Usage

showGroup(map, group)

hideGroup(map, group)

Arguments

map
the map to modify

group
character vector of one or more group names to show or hide

See Also

addLayersControl to allow users to show/hide layer groups interactively

[Package leaflet version 2.0.1 Index]