validateCoords {leaflet}	R Documentation

Utility function to check if a coordinates is valid

Description

Utility function to check if a coordinates is valid

Usage

validateCoords(lng, lat, funcName, warn = TRUE, mode = c("point",
  "polygon"))

Arguments

lng
vector with longitude values

lat
vector with latitude values

funcName
Name of calling function

warn
A boolean. Whether to generate a warning message if there are rows with missing/invalid data

mode
if "point" then warn about any NA lng/lat values; if "polygon" then NA values are expected to be used as polygon delimiters

[Package leaflet version 2.0.1 Index]