Add a tile layer from a known map provider

Add a tile layer from a known map provider

Usage

addProviderTiles(
  map,
  provider,
  layerId = NULL,
  group = NULL,
  options = providerTileOptions(),
  check = TRUE
)

providerTileOptions(
  errorTileUrl = "",
  noWrap = FALSE,
  opacity = NULL,
  zIndex = NULL,
  updateWhenIdle = NULL,
  detectRetina = FALSE,
  ...
)

Arguments

map

the map to add the tile layer to

provider

the name of the provider (see https://leaflet-extras.github.io/leaflet-providers/preview/ and https://github.com/leaflet-extras/leaflet-providers)

layerId

the layer id to assign

group

the name of the group the newly created layers should belong to (for clearGroup() and addLayersControl() purposes). Human-friendly group names are permitted–they need not be short, identifier-style names.

options

tile options

check

Check that the specified provider matches the available currently loaded leaflet providers? Defaults to TRUE, but can be toggled to FALSE for advanced users.

errorTileUrl, noWrap, opacity, zIndex, updateWhenIdle, detectRetina

the tile layer options; see https://web.archive.org/web/20220702182250/https://leafletjs.com/reference-1.3.4.html#tilelayer

...

named parameters to add to the options

Value

modified map object

Examples

leaflet() %>%
  addProviderTiles("Esri.WorldTopoMap") %>%
  addProviderTiles("CartoDB.DarkMatter")