Add a simple Graticule on the map see https://github.com/ablakey/Leaflet.SimpleGraticule

Add a simple Graticule on the map see https://github.com/ablakey/Leaflet.SimpleGraticule

Usage

addSimpleGraticule(
  map,
  interval = 20,
  showOriginLabel = TRUE,
  redraw = "move",
  hidden = FALSE,
  zoomIntervals = list(),
  layerId = NULL,
  group = NULL
)

Arguments

map

a map widget object

interval

The spacing in map units between horizontal and vertical lines.

showOriginLabel

true Whether or not to show '(0,0)' at the origin.

redraw

on which map event to redraw the graticule. On move is default but moveend can be smoother.

hidden

hide on start

zoomIntervals

use different intervals in different zoom levels. If not specified, all zoom levels use value in interval option.

layerId

the layer id

group

the name of the group this layer belongs to.

Examples

leaflet() %>%
  addTiles() %>%
  addSimpleGraticule()